fix: update script
