--- scripts/components/scroll-to-section.js
@@ -1,12 +1,0 @@
-export function scrollToSection() {
-  const anchors = document.querySelectorAll('.js-section-link')
-
-  anchors.forEach((anchor) => {
-    anchor.addEventListener('click', function (e) {
-      e.preventDefault();
-      document.querySelector(this.getAttribute('href')).scrollIntoView({
-        behavior: 'smooth'
-      });
-    });
-  })
-}
