--- scripts/components/year.js
@@ -1,5 +1,0 @@
-export function setYearInFooter() {
-  const yearContainer = document.querySelector('.js-year')
-  const year = new Date().getFullYear()
-  yearContainer.textContent = year
-}
